requirements.txt
django==4.0.4
djangorestframework==3.13.1
mysqlclient==2.1.0
django-elasticsearch-dsl==7.2.2    
elastic-transport==8.1.2    
elasticsearch==7.17.3   
elasticsearch-dsl==7.4.0 